BiNormalDistributions

Documentation for BiNormalDistributions.

Index

BiNormalDistributions.BiNormal

BiNormalDistributions.componentcdfs

BiNormalDistributions.componentpdfs

BiNormalDistributions.kdemaxes

BiNormalDistributions.maxes

BiNormalDistributions.∇loglikelihood

Distributions.cdf

Distributions.cf

Distributions.mgf

Distributions.pdf

Statistics.mean

Statistics.median

Statistics.quantile

Statistics.var

StatsAPI.loglikelihood

StatsAPI.params

StatsBase.entropy

StatsBase.kurtosis

StatsBase.moment

StatsBase.moment

StatsBase.skewness

Types reference

BiNormalDistributions.BiNormal — Type

BiNormal{T<:Real,W<:Real} <: ContinuousUnivariateDistribution
BiNormal(λ::W, N₁::Normal{T}, N₂::Normal{T})
BiNormal(λ::W, μ₁::T, σ₁::T, μ₂::T, σ₂::T)

Linear combination of two Gaussians.

The distribution is defined by five parameters:

λ: weight of the first normal distribution N₁. Since the N₁ is presumed to have more influence, λ must be within 1/2 and 1.

μ₁: mean of the first normal distribution N₁.

σ₁: variance of the first normal distribution N₁.

μ₂: mean of the second normal distribution N₂.

σ₂: variance of the second normal distribution N₂.

Domain transformation:

If the weight (λ) is available as a real number instead, the following domain transformation can be used:

\[β = \ln \frac{2λ - 1}{2 - 2λ}, \quad λ = \frac{2 + e^{-β}}{2 + 2e^{-β}}\]

Here, $λ ∈ [1/2, 1]$ and $β ∈ ℝ$.

Functions reference

BiNormalDistributions.componentcdfs — Method

componentcdfs(d::BiNormal, x::Real)

Return the (weighted) component pdfs of the BiNormal distribution, that is, $λ F_N(x; μ_1, σ_1)$ and $(1-λ) F_N(x; μ_2, σ_2)$, where $F_N$ is the cdf of the normal distribution.

BiNormalDistributions.componentpdfs — Method

componentpdfs(d::BiNormal, x::Real)

Return the (weighted) component pdfs of the BiNormal distribution, that is, $λ N(x; μ_1, σ_1)$ and $(1-λ) N(x; μ_2, σ_2)$, where $N$ is the pdf of the normal distribution.

BiNormalDistributions.kdemaxes — Function

kdemaxes(x, n = nothing)

Find the most prominent occurences in a data series. Calculated by making a kernel density estimate and then finding the peaks of the KDE when treated as a signal.

Returns the interpolated KDE and the maxima of the KDE.

See also: maxes.

BiNormalDistributions.maxes — Function

maxes(x, n = nothing)

Find the n most prominent maxima of a signal/curve x.

BiNormalDistributions.∇loglikelihood — Method

∇loglikelihood(d::BiNormal, x)

Gradient of the loglikelihood function with respect to the parameters of d.

Distributions.cdf — Method

cdf(d::BiNormal, x::Real)

The cumulative density function (cdf) is

\[F(x; λ, μ_1, σ_1, μ_2, σ_2) = λ F_N(x; μ_1, σ_1) + (1-λ) F_N(x; μ_2, σ_2)\]

where $F_N$ is the cdf of the normal distribution.

Distributions.cf — Method

cf(d::BiNormal, t)

Characteristic function ($φ_d$) of a bi-normal distribution d. The mathematical definition is:

\[φ_d(t) = λ \exp\left(itμ_1 - \tfrac{1}{2} t^2 σ_1^2\right) + (1-λ) \exp\left(itμ_2 - \tfrac{1}{2} t^2 σ_2^2\right)\]

Distributions.mgf — Method

mgf(d::BiNormal, t)

Moment generating function ($M_d$) of a bi-normal distribution d. The mathematical definition is:

\[M_d(t) = λ \exp\left(t μ_1 + \tfrac{1}{2} t^2 σ_1^2\right) + (1-λ) \exp\left(t μ_2 + \tfrac{1}{2} t^2 σ_2^2\right)\]

Distributions.pdf — Method

pdf(d::BiNormal, x::Real)

The probability density function (pdf) is

\[f(x; λ, μ_1, σ_1, μ_2, σ_2) = λ N(x; μ_1, σ_1) + (1-λ) N(x; μ_2, σ_2)\]

where $N$ is the pdf of the normal distribution.

Statistics.mean — Method

mean(d::BiNormal)

Mean of the bi-normal distribution is $μ = λ μ_1 + (1 - λ) μ_2$.

Statistics.median — Method

median(d::BiNormal)

Find the median of a BiNormal distribution d. Note that this doesn't have an analytical solution, so a root-finding algorithm from Roots.jl is employed.

Statistics.quantile — Method

quantile(d::BiNormal, q::Real)

Use Newton's method to find the quantile for BiNormal distribution d.

Statistics.var — Method

var(d::BiNormal)

Variance of the bi-normal distribution is

\[\begin{align*} σ^2 &= λσ_1^2 + (1-λ)σ_2^2 + λ(1-λ) (μ_1 - μ_2)^2 \\ &= λ (σ_1^2 + μ_1^2) + (1 - λ) (σ_2^2 + μ_2^2) - μ^2 \end{align*}\]

StatsAPI.loglikelihood — Method

loglikelihood(d::BiNormal, x)

Get the total log-likelihood of a BiNormal distribution d producing samples x.

StatsAPI.params — Method

params(d::BiNormal)

Return the parameters of a BiNormal distribution: $(λ, μ_1, σ_1, μ_2, σ_2)$.

StatsBase.entropy — Method

entropy(d::BiNormal)

Calculate the entropy of a BiNormal distribution d, evaluated numerically.

StatsBase.kurtosis — Method

kurtosis(d::BiNormal)

Mathematical definition:

\[\frac{ λ (μ_1^4 + 3 σ_1^4 + 6 μ_1^2 σ_1^2) + (1-λ) (μ_2^4 + 3 σ_2^4 + 6 μ_2^2 σ_2^2) + 3 μ^2 (μ^2 + 2 σ^2) - 4 μ [λ μ_1 (μ_1^2 + 3 σ_1^2) + (1-λ) μ_2 (μ_2^2 + 3 σ_2^2)] }{ σ^4 }\]

where $μ$ is the mean of d and $σ$ is the standard deviation of d

StatsBase.moment — Method

moment(d::BiNormal, k)

Compute the k-th moment of the BiNormal distribution around 0.

StatsBase.moment — Method

moment(d::Normal, Val(k))

Compute the k-th moment of the normal distribution around 0.

StatsBase.skewness — Method

skewness(d::BiNormal)

Mathematical definition:

\[γ = \frac{λ μ_1 (μ_1^2 + σ_1^2) + (1 - λ) μ_2 (μ_2^2 + σ_2^2) - μ (3 σ^2 + μ^2)}{σ^3}\]

where $μ$ is the mean of d and $σ$ is the standard deviation of d